requirements.txt
boto3
-e .
attrs==19.3.0
aws-cdk.assets==1.18.0
aws-cdk.aws-apigateway==1.18.0
aws-cdk.aws-applicationautoscaling==1.18.0
aws-cdk.aws-autoscaling==1.18.0
aws-cdk.aws-autoscaling-common==1.18.0
aws-cdk.aws-autoscaling-hooktargets==1.18.0
aws-cdk.aws-certificatemanager==1.18.0
aws-cdk.aws-cloudformation==1.18.0
aws-cdk.aws-cloudfront==1.18.0
aws-cdk.aws-cloudwatch==1.18.0
aws-cdk.aws-codebuild==1.18.0
aws-cdk.aws-codecommit==1.18.0
aws-cdk.aws-codepipeline==1.18.0
aws-cdk.aws-dynamodb==1.18.0
aws-cdk.aws-ec2==1.18.0
aws-cdk.aws-ecr==1.18.0
aws-cdk.aws-ecr-assets==1.18.0
aws-cdk.aws-ecs==1.18.0
aws-cdk.aws-elasticloadbalancing==1.18.0
aws-cdk.aws-elasticloadbalancingv2==1.18.0
aws-cdk.aws-events==1.18.0
aws-cdk.aws-events-targets==1.18.0
aws-cdk.aws-iam==1.18.0
aws-cdk.aws-kms==1.18.0
aws-cdk.aws-lambda==1.18.0
aws-cdk.aws-logs==1.18.0
aws-cdk.aws-route53==1.18.0
aws-cdk.aws-route53-targets==1.18.0
aws-cdk.aws-s3==1.18.0
aws-cdk.aws-s3-assets==1.18.0
aws-cdk.aws-secretsmanager==1.18.0
aws-cdk.aws-servicediscovery==1.18.0
aws-cdk.aws-sns==1.18.0
aws-cdk.aws-sns-subscriptions==1.18.0
aws-cdk.aws-sqs==1.18.0
aws-cdk.aws-ssm==1.18.0
aws-cdk.aws-stepfunctions==1.18.0
aws-cdk.core==1.18.0
aws-cdk.cx-api==1.18.0
aws-cdk.region-info==1.18.0
cattrs==0.9.0
importlib-resources==1.0.2
jsii==0.20.8
publication==0.0.3
python-dateutil==2.8.1
six==1.13.0
typing-extensions==3.7.4.1
